refactor: separate pet request
